b: fix circular import
