+++ data/transforms/vision/classification.py
@@ -5,7 +5,7 @@ import warnings
 import torchvision
 import torchvision.transforms.functional as TF
 from data.transforms import _BaseTransform
-from data.transforms.vision import str2interpolation
+from . import str2interpolation
 from torchvision.transforms.functional import InterpolationMode
 
 _RANDOM_INTERPOLATION = (str2interpolation("bilinear"), str2interpolation("bicubic"))
